Remove unused store action and centralize hex HSL conversion
