+++ src/utils/presets.ts
@@ -1,3 +1,5 @@
+import { hexToHsl } from './canvasColorUtils';
+
 export interface Preset {
   id: string;
   name: string;
@@ -6,32 +8,6 @@ export interface Preset {
   strokeOpacity: number;
   fillColor: string;
   fillOpacity: number;
-}
-
-// Helper function to convert hex to HSL
-function hexToHsl(hex: string): [number, number, number] {
-  const r = parseInt(hex.slice(1, 3), 16) / 255;
-  const g = parseInt(hex.slice(3, 5), 16) / 255;
-  const b = parseInt(hex.slice(5, 7), 16) / 255;
-
-  const max = Math.max(r, g, b);
-  const min = Math.min(r, g, b);
-  let h = 0;
-  let s = 0;
-  const l = (max + min) / 2;
-
-  if (max !== min) {
-    const d = max - min;
-    s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
-    switch (max) {
-      case r: h = (g - b) / d + (g < b ? 6 : 0); break;
-      case g: h = (b - r) / d + 2; break;
-      case b: h = (r - g) / d + 4; break;
-    }
-    h /= 6;
-  }
-
-  return [h * 360, s * 100, l * 100];
 }
 
 // Get the primary color for sorting (stroke if available, otherwise fill)
@@ -144,8 +120,8 @@ export const PRESETS: Preset[] = [
     if (colorB === '#808080' && colorA !== '#808080') return -1;
     if (colorA === '#808080' && colorB === '#808080') return 0;
 
-    const [hueA] = hexToHsl(colorA);
-    const [hueB] = hexToHsl(colorB);
+    const { h: hueA } = hexToHsl(colorA);
+    const { h: hueB } = hexToHsl(colorB);
 
     return hueA - hueB;
   })
